+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.0] - 2020-11-11
+
+### Added
+
+- Context manager support for Model and Utility classes
+- pilsner.Utility.ignore_node() method to arbitrary ignore labels present in the model
+- pilsner.Model instance created with `simple=True` parameter does not store attributes or otherwise interact with a database
+
 ## [0.0.1] - 2020-10-08
 
 ### Added